FIX: uint32_t to uint64_t FIX: Breakpoint prible thus variable size

+++ src/convert.c
@@ -1,14 +1,14 @@
 #include "convert.h"
 
 
-uint32_t str_to_hex(const char* str)
+uint64_t str_to_hex(const char* str)
 {
 
 	return (uint64_t) strtol(str, NULL, 16);
 }
 
 
-uint32_t hex_to_str(uint32_t hex)
+uint64_t hex_to_str(uint64_t hex)
 {
 	return 10;
 }
